src/auth/requirements.txt
astroid==3.0.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dill==0.3.7
Flask==3.0.0
Flask-MySQLdb==2.0.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mysqlclient==2.2.0
parso==0.8.3
platformdirs==3.11.0
PyJWT==2.8.0
pylint==3.0.2
python-decouple==3.8
tomli==2.0.1
tomlkit==0.12.2
typing_extensions==4.8.0
Werkzeug==3.0.1
